+++ src/classes/interactions-handler.js
