+++ src/events/create-interactions.js
@@ -1,14 +1,14 @@
 const { Events, Interaction } = require('discord.js');
-const InteractionsHandler = require('../classes/InteractionsHandler.js');
+const interactionsHandler = require('../classes/interactions-handler.js');
 
 module.exports = {
     name: Events.InteractionCreate,
     /** @param { Interaction } interaction  */
     async execute(interaction) {
         if (interaction.isChatInputCommand()) {
-            InteractionsHandler.handleSlashcommands(interaction);
+            interactionsHandler.handleSlashcommands(interaction);
         } else if (interaction.isAutocomplete()) {
-            InteractionsHandler.handleAutocomplete(interaction);
+            interactionsHandler.handleAutocomplete(interaction);
         };
     },
 };
